+++ migrations/mytbdb/sqls/sqlite3/20221010202000-add-poi-down.sql
@@ -1,0 +1,7 @@
+-- ----------------
+-- add poi-tables
+-- ----------------
+
+DROP table if exists poi_info;
+DROP table if exists poi_keyword;
+DROP table if exists poi;