+++ src/shared/tdoc-commons/model/repository/sql-mytbdb-news.config.ts
@@ -310,7 +310,7 @@ export class SqlMytbDbNewsConfig {
                 noFacet: true
             },
             'type_txt': {
-                constValues: ['trip', 'location', 'track', 'route', 'image', 'odimgobject', 'video', 'news', 'destination', 'info', 'playlist'],
+                constValues: ['trip', 'location', 'track', 'route', 'image', 'odimgobject', 'video', 'news', 'destination', 'info', 'playlist', 'poi'],
                 filterField: '"news"',
                 selectLimit: 1
             },
@@ -388,6 +388,8 @@ export class SqlMytbDbNewsConfig {
             info_id_is: '"666dummy999"',
             video_id_is: '"666dummy999"',
             video_id_i: '"666dummy999"',
+            poi_id_i: '"666dummy999"',
+            poi_id_is: '"666dummy999"',
             track_id_i: '"666dummy999"',
             track_id_is: '"666dummy999"',
             trip_id_i: '"666dummy999"',